put decode function in decode.go, NullTime type in time.go
